+++ src/SampleSimilarity/main.cpp
@@ -88,7 +88,7 @@ public:
 			}
 			else if(mode=="gsvar")
 			{
-				genotype_data << (use_roi ? SampleSimilarity::genotypesFromGSvar(filename, include_gonosomes, skip_multi, roi_reg) : SampleSimilarity::genotypesFromGSvar(filename, include_gonosomes, skip_multi));
+				genotype_data << (use_roi ? SampleSimilarity::genotypesFromGSvar(filename, include_gonosomes, roi_reg) : SampleSimilarity::genotypesFromGSvar(filename, include_gonosomes));
 			}
 			else
 			{